fixed the image, added some visual tweeks
line 6: here you should be more specific that just 'div' div#busCard or just #busCard is good too.  Otherwise just div will apply the css to all divs on the page
The reason the image was being clobbered is because you were replacing the entire contents of 'busCard' with the name etc.  the image was IN the busCard block so it got clobbered(written over)  I broke the image and text into multiple areas.