changing image
i originally had the smiley face in there to test if an image would work, now i pulled in a more appropriate image from the internet to use on the business card.  I think it looks perfect now and is complete.  Let me know if you see anything else that needs to be fixed.  Thanks!